+++ src/control.entity.ts
@@ -31,7 +31,7 @@ import { ClusterId, ClusterRegistry } from 'matterbridge/matter/types';
 
 import { HassEntity, HassState } from './homeAssistant.js';
 import { MutableDevice } from './mutableDevice.js';
-import { hassDomainConverter, hassCommandConverter, hassSubscribeConverter } from './converters.js';
+import { hassDomainConverter, hassCommandConverter, hassSubscribeConverter, miredsToKelvin, kelvinToMireds } from './converters.js';
 
 /**
  * Look for supported binary_sensors of the current entity
@@ -107,13 +107,16 @@ export function addControlEntity(
   // Configure the ColorControl cluster default values and features.
   // prettier-ignore
   if (domain === 'light' && (mutableDevice.get(endpointName).deviceTypes.includes(colorTemperatureLight) || mutableDevice.get(endpointName).deviceTypes.includes(extendedColorLight))) {
-    log.debug(`= colorControl device ${CYAN}${entity.entity_id}${db} supported_color_modes: ${CYAN}${state.attributes['supported_color_modes']}${db} min_mireds: ${CYAN}${state.attributes['min_mireds']}${db} max_mireds: ${CYAN}${state.attributes['max_mireds']}${db}`);
+    log.debug(`= colorControl device ${CYAN}${entity.entity_id}${db} supported_color_modes: ${CYAN}${state.attributes['supported_color_modes']}${db} min_color_temp_kelvin: ${CYAN}${state.attributes['min_color_temp_kelvin']}${db} max_color_temp_kelvin: ${CYAN}${state.attributes['max_color_temp_kelvin']}${db}`);
+    const minMireds = kelvinToMireds(state.attributes['max_color_temp_kelvin'] ?? miredsToKelvin(500, 'floor'), 'floor');
+    const maxMireds = kelvinToMireds(state.attributes['min_color_temp_kelvin'] ?? miredsToKelvin(147, 'floor'), 'floor');
+    log.debug(`= colorControl device ${CYAN}${entity.entity_id}${db} supported_color_modes: ${CYAN}${state.attributes['supported_color_modes']}${db} min_mireds: ${CYAN}${minMireds}${db} max_mireds: ${CYAN}${maxMireds}${db}`);
     if (isValidArray(state.attributes['supported_color_modes']) && !state.attributes['supported_color_modes'].includes('xy') && !state.attributes['supported_color_modes'].includes('hs') && !state.attributes['supported_color_modes'].includes('rgb') &&
       !state.attributes['supported_color_modes'].includes('rgbw') && !state.attributes['supported_color_modes'].includes('rgbww') && state.attributes['supported_color_modes'].includes('color_temp')
     ) {
-      mutableDevice.addClusterServerColorTemperatureColorControl(endpointName, state.attributes['color_temp'] ?? 250, state.attributes['min_mireds'] ?? 147, state.attributes['max_mireds'] ?? 500);
+      mutableDevice.addClusterServerColorTemperatureColorControl(endpointName, minMireds, maxMireds);
     } else {
-      mutableDevice.addClusterServerColorControl(endpointName, state.attributes['color_temp'] ?? 250, state.attributes['min_mireds'] ?? 147, state.attributes['max_mireds'] ?? 500);
+      mutableDevice.addClusterServerColorControl(endpointName, minMireds, maxMireds);
     }
   }